backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
pydantic==2.5.0
sqlalchemy==2.0.23
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.0
boto3==1.34.0
Pillow==10.1.0
